+++ src/jeap_pipeline/__init__.py
@@ -1,6 +1,5 @@
 # __init__.py
-from .ecs_deployment_checker import wait_until_new_deployment_has_occurred, wait_until_deployments_completed, \
-    get_deployment_status, DeploymentStatus
+from .ecs_deployment_checker import wait_until_deployments_completed, get_deployment_status, DeploymentStatus
 from .ecs_undeployment_checker import wait_until_undeployment_has_finished, is_service_undeployed
 from .github_dispatch_event import send_dispatch_event
 from .pact_pacticipants import get_pacticipant_names